requirements.txt
psycopg2-binary
sqlalchemy
pandas
requests
pdfplumber
pdf2image
paddleocr
paddlepaddle
reportlab
python-dateutil
schedule
numpy
flask
plotly
google-auth
google-auth-oauthlib
google-api-python-client
openpyxl
python-dotenv
